JDK support for hermetic JAR packaged jimage (aka runtime image).
This handles the JDK runtime image that's packaged within the hermetic JAR at a specific file offset (page aligned). In hermetic JAR, the <JDK>/lib/'modules' data is packed within the JVM data section, which is between the ELF section and the JAR section in a deploy JAR. The start offset of the 'modules' data is page aligned. Special launcher argument that can be passed to the VM:

  -XX:UseHermeticJDK=<deploy_jar_path>,<jimage_start_offset>

The jimage start offset is used when opening/mapping and reading the jimage.