Support hermetic JAR packaged lib/ct.sym file.
Removed the use of the static final 'symbolFileLocation' field, which defines the path subcomponents of the 'lib/ct.sym' file. Instead, the path elements are passed as the arguments of JavaHome.getJDKResource(). That provides more complete and precise logging information for verifying the runtime access of ct.sym file path.

This change requires JavaHome class in the boot JDK.