Add DisableHermetic flag for disabling hermetic Java runtime.
If a hermetic JAR is used, setting DisableHermetic to true disables the hermetic Java runtime (the JDK modules image and JDK resources) packaged in JAR. The flag has no effect, if the JAR is non-hermetic. It does not affect the JDK static support either.

This option should only be used for a few limited use cases, such as:
- During testing/debugging for isolating JDK issues when necessary.