Support hermetic packaged lib/security/cacerts in sun.security.ssl.TrustStoreManager$TrustStoreDescriptor and sun.security.util.AnchorCertificates.
Use jdk.internal.misc.JavaHome to access JDK default store, lib/security/cacerts in TrustStoreDescriptor. If 'javax.net.ssl.trustStore' property is set and the value is not 'NONE', the specified store is accessed as a regular file using Path.of() API. The original semantics should not be affected by the change.